restructure caching
