+++ internal/cache/context.go
@@ -1,4 +1,4 @@
-package gbimporter
+package cache
 
 import "go/build"
