+++ internal/cache/samepath_unix.go
@@ -1,12 +1,8 @@
-// +build windows
+// +build !windows
 
-package gbimporter
-
-import (
-	"strings"
-)
+package cache
 
 // samePath checks two file paths for their equality based on the current filesystem
 func samePath(a, b string) bool {
-	return strings.EqualFold(a, b)
+	return a == b
 }
