+++ internal/cache/samepath_windows.go
@@ -1,8 +1,12 @@
-// +build !windows
+// +build windows
 
-package gbimporter
+package cache
+
+import (
+	"strings"
+)
 
 // samePath checks two file paths for their equality based on the current filesystem
 func samePath(a, b string) bool {
-	return a == b
+	return strings.EqualFold(a, b)
 }
